Feat [#5] Global 패키지 생성
SuccessResponse Generic으로 수정
