+++ jumpit/src/main/java/org/sopt/jumpit/global/config/JpaAuditingConfig.java
@@ -1,4 +1,4 @@
-package org.sopt.jumpit.config;
+package org.sopt.jumpit.global.config;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
