+++ jumpit/src/main/java/org/sopt/jumpit/global/exception/NotFoundException.java
@@ -1,6 +1,6 @@
-package org.sopt.jumpit.exception;
+package org.sopt.jumpit.global.exception;
 
-import org.sopt.jumpit.common.dto.message.ErrorMessage;
+import org.sopt.jumpit.global.common.dto.message.ErrorMessage;
 
 public class NotFoundException extends BusinessException {
     public NotFoundException(ErrorMessage errorMessage) {
